requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-JWT-Extended>=4.4
Flask-Mail>=0.9
Flask-SQLAlchemy>=3.0
requests>=2.0
itsdangerous>=2.0
Werkzeug>=2.0
Pillow>=9.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
opencv-python>=4.5.0
